[thin_ls (rust)] Improve efficiency by reading shared nodes only once
A data block is treated as shared if one of its ancestors
along the path is shared. In other words, all the data blocks
below a node are shared if the node has ref count greater
than one. Therefore, the number of shared data blocks of
a device could be determined by using the stored ref counts
and statistics gathered from its leaves, which eliminates
the need of recursing to the shared leaves.