[thin_dump] Prompt for repair on input errors
thin_dump should display the repairing hint if there's any error in the
input metadata rather than the output process. A context object thus is
added to the returned error for callers to identify the error type.

Note that a broken pipe error (EPIPE) is returned as an output error since
the Rust std library ignores SIGPIPE by default [1][2]. However, due to
the dump process leverages NodeVisitor, a low-level output error is
stringified into a BTreeError, and thus the caller cannot tell whether
it's broken pipe or not.

[1] https://github.com/rust-lang/rust/pull/13158
[2] https://github.com/rust-lang/rust/issues/62569