[cache_dump] Terminate the process immediately on error
Previously, cache_dump uses BTreeWalker to visit the mappings and hint
arrays recursively, but there are drawbacks to this approach:

- The BTreeWalker was designed for *_check but not *_dump applications.
  It does not stop immediately when an error happens, neither a checksum
  error nor a broken pipe error (EPIPE).
- It's not easy to describe complex errors using the types returned by
  BTreeWalker. Both BTreeError and ArrayError should be used to describe
  errors on the data structures themself rather than post-processing,
  such as dump or restore.

An iterative way is adopted for the purposes above. It simply loops over
the collected array blocks. This way simplifies error handling and
avoids nested encapsulation between ArrayError and BTreeError.

Also, display the repairing hints on metadata error.