[thin_ls] Employ the streaming architecture to accelerate scanning
thin_ls was slow due to we've moved to direct io and thus the
single-threaded second pass became the performance bottleneck. A 16GB
metadata might spent 8 mins or more to finish the scanning, which is
unacceptable compared to that of thin_check.

In this commit, we adopt the streaming approach used in thin_check to
accelerate the scanning of shared/exclusive mappings. The workflow is
similar, except that thin_ls performs an extra pass on scanning
exclusive leaves to count the number of shared data mappings within
after the data space map becomes available. Now thin_ls is able to
finish the 16GB metadata in 26 secs.