[file_utils] Fix the BLKGETSIZE64 ioctl for 32-bit platoforms
The nix::ioctl_read! binding macro doesn't work with the BLKGETSIZE64
ioctl because the type of the output parameter doesn't match that of the
request code. Specifically, the ioctl returns a 64-bit value on all
platforms, but the type size in the request code varies between 32-bit
and 64-bit platforms due to its reliance on the size_t type.

While this issue can be solved by using the nix::ioctl_read_bad! macro,
the solution may be less intuitive. Therefore, we use libc::ioctl
directly.