[thin_migrate] Fix file size check on the output file
- Fix potential TOCTOU issue on retrieving the output device size.
- Remove file size check on regular file output due to design
  constraints. Reasons stated below:

To avoid the TOCTOU issue, it's recommended to perform checks on an
opened file descriptor. Since thin_migrate implicitly creates the
output file with O_CREAT, checking the size of a regular file opened
this way becomes impractical. We cannot distinguish between a newly
created zero-length file and an existing empty file, in which the
latter one is unacceptable. Therefore, remove the file size check
accordingly.

blk-archive is free from this issue as there are separated command line
options for opening or creating the output file explicitly.